Convert durationToSeconds to a DurationFormatter mixin method.
