+++ timestrap/static_src/scripts/utils.js
@@ -8,17 +8,6 @@ function pad(num) {
 }
 
 
-// Convert a decimal duration (0.0) duration to a number (0) of seconds.
-function durationToSeconds(duration) {
-    if (typeof(duration) === 'number') {
-        let hours = Math.floor(duration);
-        let minutes = Math.round((duration - hours) * 60);
-        duration = hours * 3600 + minutes * 60;
-    }
-    return duration;
-}
-
-
 // Convert a number (0) of seconds to a string (0:00).
 function secondsToDurationString(duration) {
     if (typeof(duration) === 'number') {
